plate_analysis/cloud/event_router/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-tasks==2.*
